requirements.txt
cffi==1.15.1
click==8.1.3
cryptography==39.0.0
dlib==19.24.0
face-recognition==1.3.0
face-recognition-models==0.3.0
M2Crypto==0.38.0
numpy==1.23.2
opencv-python==4.6.0.66
pgi==0.0.11.2
Pillow==9.2.0
pycparser==2.21
pycryptodome==3.10.1
pyOpenSSL==23.0.0
pyscard==2.0.0
python-dateutil==2.8.2
six==1.16.0
